dockerclient: tighten up cleanup logic
When cleaning up at the end of a build, instead of iterating through
stages and running their cleanup logic in whatever order a map of the
stages gives them to us, run through them in the exact reverse of the
order in which they were executed.

Reorder the cleanups done for each stage so that if the stage's result
was committed to create an image, we try to remove that image before
removing the stage's container, and if that container was based on a
scratch image, we remove the scratch image after removing the container.

Removing an image produced from an intermediate stage will fail if that
stage's image is an ancestor of the final image, so ignore errors we
encounter while trying to remove those images.  Treat every other error
we encounter while trying to remove a container or an image during
conformance tests as a failure in the test.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>