dockerclient: COPY --from should ignore the source's WORKDIR
When COPY is given --from=${stage|image}, the stage's or image's WORKDIR
should not be incorporated into the path where we go to look for the
source when the source is not specified with an absolute path name.
We previously incorporated the WORKDIR value, but that is not compatible
with what `docker build` does.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>